requirements.txt
psycopg2-binary==2.9.9
psycopg2==2.9.9
python-dotenv==1.0.1
pytest~=8.3.3